spotify/analysis/requirements.txt
python-decouple
flask
psycopg2-binary

